+++ code/CMakeLists.txt
@@ -8,7 +8,7 @@ set(PROJECT_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/src/main.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/gfx_helper.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/gui/gui_main.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/src/window/gfx_window.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/gfx/gfx_window.cpp
 )
 
 set(PROJECT_INCLUDES
@@ -16,7 +16,7 @@ set(PROJECT_INCLUDES
     ${CMAKE_CURRENT_SOURCE_DIR}/inc/worldweaver/gfx_helper.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/inc/worldweaver/star.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/inc/gui/gui_main.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/inc/window/gfx_window.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/inc/gfx/gfx_window.hpp
     )
 
 add_executable(worldweaver
