+++ code/inc/gfx/base_window.hpp
